remove modeling as a dep for data
