+++ asapdiscovery-data/asapdiscovery/data/backend/openeye.py
@@ -1,4 +1,5 @@
 from base64 import b64decode, b64encode
+from functools import reduce
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Union  # noqa: F401
 from warnings import warn
@@ -16,6 +17,9 @@ from openeye import (  # noqa: F401
     oespruce,
     oeszybki,
 )
+
+from asapdiscovery.modeling.modeling import trim_small_chains
+from schema.schema_base import MoleculeFilter
 
 # exec on module import
 
@@ -1123,3 +1127,213 @@ def bytes64_to_oedu(bytes: bytes) -> oechem.OEDesignUnit:
     if not retcode:
         oechem.OEThrow.Fatal("Cannot read DesignUnit from bytes")
     return du
+
+
+def split_openeye_mol(
+    complex_mol,
+    molecule_filter: Optional[Union[str, list[str], MoleculeFilter]] = None,
+    prot_cutoff_len=10,
+    keep_one_lig=True,
+) -> dict:
+    """
+    Split an OpenEye-loaded molecule into protein, ligand, etc.
+    Uses the OpenEye OESplitMolComplex function, which automatically splits out
+    only the first ligand binding site it sees.
+
+    Parameters
+    ----------
+    complex_mol : oechem.OEMolBase
+        Complex molecule to split.
+    molecule_filter : MoleculeFilter
+        Molecule filter object that contains the filter criteria.
+    prot_cutoff_len : int, default=10
+        Minimum number of residues in a protein chain required in order to keep
+    keep_one_lig : bool, default=True
+        Only keep one copy of the ligand in lig_mol
+
+    Returns
+    -------
+    dict
+        A dict containing the protein ("prot"), ligand ("lig"), water atoms ("wat"),
+        and all other atoms ("oth")
+    """
+    # These are the objects that will be returned at the end
+    lig_mol = oechem.OEGraphMol()
+    prot_mol = oechem.OEGraphMol()
+    water_mol = oechem.OEGraphMol()
+    oth_mol = oechem.OEGraphMol()
+
+    if molecule_filter is None:
+        molecule_filter = MoleculeFilter()
+    elif type(molecule_filter) is str:
+        molecule_filter = MoleculeFilter(components_to_keep=[molecule_filter])
+    elif type(molecule_filter) is list:
+        molecule_filter = MoleculeFilter(components_to_keep=molecule_filter)
+    else:
+        molecule_filter = molecule_filter
+
+    # Make splitting split out covalent ligands possible
+    # TODO: look into different covalent-related options here
+    opts = oechem.OESplitMolComplexOptions()
+    opts.SetSplitCovalent(True)
+    opts.SetSplitCovalentCofactors(True)
+
+    # Protein splitting options
+    # Default of all atoms identified as protein + complex
+    prot_only = oechem.OEMolComplexFilterFactory(
+        oechem.OEMolComplexFilterCategory_ProtComplex
+    )
+    # add in peptides as well, sometimes proteins are misidentified as peptides if bound or short
+    peptide = oechem.OEMolComplexFilterFactory(
+        oechem.OEMolComplexFilterCategory_Peptide
+    )
+
+    # combine protein and peptide filters
+    prot_only = oechem.OEOrRoleSet(prot_only, peptide)
+
+    # If protein_chains are specified, only take protein atoms from those chains
+    if len(molecule_filter.protein_chains) > 0:
+        chain_filters = [
+            oechem.OERoleMolComplexFilterFactory(
+                oechem.OEMolComplexChainRoleFactory(chain)
+            )
+            for chain in molecule_filter.protein_chains
+        ]
+
+        chain_filter = reduce(oechem.OEOrRoleSet, chain_filters)
+        prot_filter = oechem.OEAndRoleSet(prot_only, chain_filter)
+    else:
+        prot_filter = prot_only
+    opts.SetProteinFilter(prot_filter)
+
+    # Ligand splitting options
+    # Default of all atoms identified as ligand
+    lig_only = oechem.OEMolComplexFilterFactory(
+        oechem.OEMolComplexFilterCategory_Ligand
+    )
+    # If ligand_chain is specified, only take protein atoms from that chains
+    if molecule_filter.ligand_chain:
+        lig_chain = oechem.OERoleMolComplexFilterFactory(
+            oechem.OEMolComplexChainRoleFactory(molecule_filter.ligand_chain)
+        )
+        lig_filter = oechem.OEAndRoleSet(lig_only, lig_chain)
+    else:
+        lig_filter = lig_only
+    # combine with NOT peptide filter
+    lig_filter = oechem.OEAndRoleSet(lig_filter, oechem.OENotRoleSet(peptide))
+    opts.SetLigandFilter(lig_filter)
+
+    # If water_chains are specified, set up filter for them
+    if len(molecule_filter.water_chains) > 0:
+        water_only = oechem.OEMolComplexFilterFactory(
+            oechem.OEMolComplexFilterCategory_Water
+        )
+        chain_filters = [
+            oechem.OERoleMolComplexFilterFactory(
+                oechem.OEMolComplexChainRoleFactory(chain)
+            )
+            for chain in molecule_filter.water_chains
+        ]
+
+        chain_filter = reduce(oechem.OEOrRoleSet, chain_filters)
+        wat_filter = oechem.OEAndRoleSet(water_only, chain_filter)
+        # combine with NOT peptide filter
+        wat_filter = oechem.OEAndRoleSet(wat_filter, oechem.OENotRoleSet(peptide))
+        opts.SetWaterFilter(wat_filter)
+
+    # Use python 'reduce' to combine all the filters into one, otherwise OpenEye will throw an error
+    oechem.OESplitMolComplex(
+        lig_mol,
+        prot_mol,
+        water_mol,
+        oth_mol,
+        complex_mol,
+        opts,
+    )
+    # TODO: make this nicer?
+    # Get rid of any straggling extra copies of the ligand
+    prot_mol = trim_small_chains(prot_mol, prot_cutoff_len)
+
+    if molecule_filter.ligand_chain:
+        keep_lig_chain = molecule_filter.ligand_chain
+    else:
+        keep_lig_chain = None
+    # Get rid of extra copies of the ligand
+    if keep_one_lig:
+        all_lig_chains = [res.GetChainID() for res in oechem.OEGetResidues(lig_mol)]
+        # Handle the case where the input has no ligand, otherwise throws an IndexError
+        if len(all_lig_chains) > 0:
+            # Keep first alphabetically, for reproducibility
+            if not keep_lig_chain:
+                keep_lig_chain = sorted(all_lig_chains)[0]
+            for a in lig_mol.GetAtoms():
+                # Delete all atoms that don't match
+                if oechem.OEAtomGetResidue(a).GetChainID() != keep_lig_chain:
+                    lig_mol.DeleteAtom(a)
+
+    return {
+        "prot": prot_mol,
+        "lig": lig_mol,
+        "wat": water_mol,
+        "oth": oth_mol,
+        "keep_lig_chain": keep_lig_chain,
+    }
+
+
+def split_openeye_design_unit(du, lig=None, lig_title=None):
+    """
+    Parameters
+    ----------
+    du : oechem.OEDesignUnit
+        Design Unit to be saved
+    lig : oechem.OEGraphMol, optional
+
+    lig_title : str, optional
+        ID of Ligand to be saved to the Title tag in the SDF, by default None
+    include_solvent : bool, optional
+        Whether to include solvent in the complex, by default True
+
+    Returns
+    -------
+    lig : oechem.OEGraphMol
+        OE object containing ligand
+    protein : oechem.OEGraphMol
+        OE object containing only protein
+    complex : oechem.OEGraphMol
+        OE object containing ligand + protein
+    """
+    prot = oechem.OEGraphMol()
+    complex_ = oechem.OEGraphMol()
+
+    du.GetProtein(prot)
+
+    # if no ligand, return protein and complex with no ligand
+    if not du.HasLigand():
+        return None, prot, prot
+
+    if not lig:
+        lig = oechem.OEGraphMol()
+        du.GetLigand(lig)
+
+    # Set ligand title, useful for the combined sdf file
+    if lig_title:
+        lig.SetTitle(f"{lig_title}")
+
+    # Give ligand atoms their own chain "L" and set the resname to "LIG"
+    residue = oechem.OEAtomGetResidue(next(iter(lig.GetAtoms())))
+    residue.SetChainID("L")
+    residue.SetName("LIG")
+    residue.SetHetAtom(True)
+    for atom in list(lig.GetAtoms()):
+        oechem.OEAtomSetResidue(atom, residue)
+
+    # Combine protein and ligand and save
+    # TODO: consider saving water as well
+    oechem.OEAddMols(complex_, prot)
+    oechem.OEAddMols(complex_, lig)
+
+    # Clean up PDB info by re-perceiving, perserving chain ID,
+    # residue number, and residue name
+    openeye_perceive_residues(prot)
+    openeye_perceive_residues(complex_)
+    return lig, prot, complex_
