+++ asapdiscovery-modeling/asapdiscovery/modeling/tests/test_modeling.py
@@ -1,8 +1,7 @@
 import pytest
-from asapdiscovery.data.backend.openeye import load_openeye_pdb, oechem
+from asapdiscovery.data.backend.openeye import load_openeye_pdb, oechem, split_openeye_mol
 from asapdiscovery.data.testing.test_resources import fetch_test_file
-from asapdiscovery.modeling.modeling import split_openeye_mol
-from asapdiscovery.modeling.schema import MoleculeFilter
+from asapdiscovery.data.schema.schema_base import MoleculeFilter
 
 
 @pytest.fixture()
