+++ devtools/conda-envs/ubuntu-latest/data.yaml
@@ -11,7 +11,6 @@ dependencies:
   - typing-extensions
 
   # drugforge-docking
-  # drugforge-modeling
   # drugforge-spectrum
 
   # Command Line Tools & Scripting
